Adaptivity

GridapDistributed.RedistributeGlue — Type

RedistributeGlue

Glue linking two distributions of the same mesh.

new_parts: Array with the new part IDs (and comms)

old_parts: Array with the old part IDs (and comms)

parts_rcv: Array with the part IDs from which each part receives

parts_snd: Array with the part IDs to which each part sends

lids_rcv : Local IDs of the entries that are received from each part

lids_snd : Local IDs of the entries that are sent to each part

old2new : Mapping of local IDs from the old to the new mesh

new2old : Mapping of local IDs from the new to the old mesh

GridapDistributed.redistribute — Method

Redistributes an DistributedDiscreteModel to optimally rebalance the loads between the processors. Returns the rebalanced model and a RedistributeGlue instance.

GridapDistributed.redistribute_cartesian — Method

redistribute_cartesian(old_model,new_ranks,new_parts)
redistribute_cartesian(old_model,pdesc::DistributedCartesianDescriptor)

Redistributes a DistributedCartesianDiscreteModel to a new set of ranks and parts. Only redistributes into a superset of the old_model ranks (i.e. towards more processors).